GDPR-Compliant-Mini-Hospital-Management-System-main/requirements.txt
streamlit>=1.37.0
pandas>=2.1.0
cryptography>=42.0.0
python-dotenv>=1.0.0
altair>=5.0.0
